Conflicts:
	mapsforge-map-awt/src/main/java/org/mapsforge/map/awt/AwtTileBitmap.java
	mapsforge-map/src/main/java/org/mapsforge/map/layer/cache/FileSystemTileCache.java
	mapsforge-map/src/main/java/org/mapsforge/map/layer/download/tilesource/OpenStreetMapMapnik.java
	mapsforge-map/src/main/java/org/mapsforge/map/layer/renderer/DatabaseRenderer.java

+++ mapsforge-map/src/main/java/org/mapsforge/map/layer/cache/FileSystemTileCache.java
@@ -106,10 +106,12 @@ public class FileSystemTileCache extends PausableThread implements TileCache {
 	private static boolean deleteDirectory(File dir) {
 		if (dir.isDirectory()) {
 			String[] children = dir.list();
-			for (int i = 0; i < children.length; i++) {
-				boolean success = deleteDirectory(new File(dir, children[i]));
-				if (!success) {
-					return false;
+			if (children != null) {
+				for (int i = 0; i < children.length; i++) {
+					boolean success = deleteDirectory(new File(dir, children[i]));
+					if (!success) {
+						return false;
+					}
 				}
 			}
 		}
